+++ source/view/schoolView.ts
